+++ packages/two_dimensional_scrollables/lib/src/table_view/table.dart
@@ -5,7 +5,10 @@
 import 'dart:collection';
 import 'dart:math' as math;
 
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
 import '../common/implementation_span.dart';
@@ -1696,22 +1699,118 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
   }
 }
 
-class _Span extends ImplementedSpan {
+class _Span
+    with Diagnosticable
+    implements HitTestTarget, MouseTrackerAnnotation {
+  double get leadingOffset => _leadingOffset;
+  late double _leadingOffset;
+
+  double get extent => _extent;
+  late double _extent;
+
+  TableSpan get configuration => _configuration!;
+  TableSpan? _configuration;
+
   bool get isPinned => _isPinned;
   late bool _isPinned;
 
-  @override
+  double get trailingOffset {
+    return leadingOffset +
+        extent +
+        configuration.padding.leading +
+        configuration.padding.trailing;
+  }
+
+  // ---- Span Management ----
+
   void update({
     required TableSpan configuration,
     required double leadingOffset,
     required double extent,
-    bool isPinned = false,
+    required bool isPinned,
   }) {
+    _leadingOffset = leadingOffset;
+    _extent = extent;
     _isPinned = isPinned;
-    super.update(
-      configuration: configuration,
-      leadingOffset: leadingOffset,
-      extent: extent,
-    );
+    if (configuration == _configuration) {
+      return;
+    }
+    _configuration = configuration;
+    // Only sync recognizers if they are in use already.
+    if (_recognizers != null) {
+      _syncRecognizers();
+    }
   }
+
+  void dispose() {
+    _disposeRecognizers();
+  }
+
+  // ---- Recognizers management ----
+
+  Map<Type, GestureRecognizer>? _recognizers;
+
+  void _syncRecognizers() {
+    if (configuration.recognizerFactories.isEmpty) {
+      _disposeRecognizers();
+      return;
+    }
+    final Map<Type, GestureRecognizer> newRecognizers =
+        <Type, GestureRecognizer>{};
+    for (final Type type in configuration.recognizerFactories.keys) {
+      assert(!newRecognizers.containsKey(type));
+      newRecognizers[type] = _recognizers?.remove(type) ??
+          configuration.recognizerFactories[type]!.constructor();
+      assert(
+        newRecognizers[type].runtimeType == type,
+        'GestureRecognizerFactory of type $type created a GestureRecognizer of '
+        'type ${newRecognizers[type].runtimeType}. The '
+        'GestureRecognizerFactory must be specialized with the type of the '
+        'class that it returns from its constructor method.',
+      );
+      configuration.recognizerFactories[type]!
+          .initializer(newRecognizers[type]!);
+    }
+    _disposeRecognizers(); // only disposes the ones that where not re-used above.
+    _recognizers = newRecognizers;
+  }
+
+  void _disposeRecognizers() {
+    if (_recognizers != null) {
+      for (final GestureRecognizer recognizer in _recognizers!.values) {
+        recognizer.dispose();
+      }
+      _recognizers = null;
+    }
+  }
+
+  // ---- HitTestTarget ----
+
+  @override
+  void handleEvent(PointerEvent event, HitTestEntry entry) {
+    if (event is PointerDownEvent &&
+        configuration.recognizerFactories.isNotEmpty) {
+      if (_recognizers == null) {
+        _syncRecognizers();
+      }
+      assert(_recognizers != null);
+      for (final GestureRecognizer recognizer in _recognizers!.values) {
+        recognizer.addPointer(event);
+      }
+    }
+  }
+
+  // ---- MouseTrackerAnnotation ----
+
+  @override
+  MouseCursor get cursor => configuration.cursor;
+
+  @override
+  PointerEnterEventListener? get onEnter => configuration.onEnter;
+
+  @override
+  PointerExitEventListener? get onExit => configuration.onExit;
+
+  @override
+  bool get validForMouseTracker => true;
 }
